+++ src/webdav/handlers/PUT.handler.ts
@@ -13,7 +13,7 @@ import { WebDavFolderService } from '../../services/webdav/webdav-folder.service
 import { ThumbnailUtils } from '../../utils/thumbnail.utils';
 import { ThumbnailService } from '../../services/thumbnail.service';
 import { FormatUtils } from '../../utils/format.utils';
-import { XMLUtils } from '../../utils/xml.utils';
+import { UploadUtils } from '../../utils/upload.utils';
 
 export class PUTRequestHandler implements WebDavMethodHandler {
   handle = async (req: Request, res: Response) => {
@@ -21,6 +21,8 @@ export class PUTRequestHandler implements WebDavMethodHandler {
     if (!contentLength || Number.isNaN(contentLength) || contentLength <= 0) {
       contentLength = 0;
     }
+
+    await UploadUtils.checkUploadSizeLimits(contentLength);
 
     const resource = await WebDavUtils.getRequestedResource(req.url);
     webdavLogger.info(`[PUT] Request received for file at ${resource.url}`);
@@ -85,39 +87,13 @@ export class PUTRequestHandler implements WebDavMethodHandler {
       const networkUploadTimer = CLIUtils.timer();
       const abortable = new AbortController();
 
-      try {
-        fileId = await networkFacade.uploadFile({
-          from: fileStream,
-          size: contentLength,
-          bucketId: bucket,
-          progressCallback,
-          abortSignal: abortable.signal,
-        });
-      } catch (e) {
-        aborted = true;
-        abortable.abort();
-        const error = e as Error;
-        if (bufferStream) {
-          req.unpipe(bufferStream);
-          bufferStream.destroy();
-        }
-        if (error.message.toLowerCase().includes('file is too big')) {
-          webdavLogger.error('[PUT] ❌ File is too big (' + FormatUtils.humanFileSize(contentLength) + ')');
-          const errorBodyXML = XMLUtils.toWebDavXML(
-            {
-              [XMLUtils.addDefaultNamespace('responsedescription')]: error.message,
-            },
-            {},
-            'error',
-          );
-          res.set('Content-Type', 'application/xml; charset="utf-8"');
-          res.status(413).send(errorBodyXML);
-          req.destroy();
-          return;
-        }
-
-        throw e;
-      }
+      fileId = await networkFacade.uploadFile({
+        from: fileStream,
+        size: contentLength,
+        bucketId: bucket,
+        progressCallback,
+        abortSignal: abortable.signal,
+      });
 
       res.on('close', async () => {
         aborted = true;
